cambios en los controladores
cambios en el controlador perosna editar

+++ src/main/java/com/example/ApiVeterinaria/Repository/IPersona.java
@@ -4,5 +4,5 @@ package com.example.ApiVeterinaria.Repository;
 import com.example.ApiVeterinaria.Modelos.Persona;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IPersona extends CrudRepository<Persona, Long> {
+public interface IPersona extends CrudRepository<Persona, Integer> {
 }
